backend/requirements.txt
flask
python-dotenv
flask_smorest
sqlalchemy
flask_sqlalchemy
passlib
flask-jwt-extended
redis
flask-cors
functools